Fix Dialyzer issues with DB seed task.
First, the run function wasn't being recognized. I thought it was valid (came from AI model recommendation) and the code was functioning so I didn't think much more of it. Another thing that needed to be fixed was updating the Slax Dialyzer configuration regarding its Persistent Lookup Table (PLT). It was missing some information on the expected behavior for Mix.Task and was unable to infer whether the task was correctly implemented or not. I extended the project's definition with mix to ensure dialyzer wouldn't be confused anymore.