Refactored seeding logic.
Prior, I had created a custom task for seeding data in Postgres. An issue that rose was that when the Docker container is up, it doesn't mean that the application is running when executing its entrypoint. From research, I saw that the correct practice was to use priv/repo/seed.exs as the task seeding the DB. Furthermore I also found the function to invoke to ensure that the app is running prior to seeding.