Listing & navigating across chat rooms (#19)
* New aliases for building the app.

* Updated the spec.

* Listing all chat rooms in sidebar menu.

* Renamed file with correct extension.

* Ensuring mix task present prior running it in docker.

* Fix Dialyzer issues with DB seed task.

First, the run function wasn't being recognized. I thought it was valid (came from AI model recommendation) and the code was functioning so I didn't think much more of it. Another thing that needed to be fixed was updating the Slax Dialyzer configuration regarding its Persistent Lookup Table (PLT). It was missing some information on the expected behavior for Mix.Task and was unable to infer whether the task was correctly implemented or not. I extended the project's definition with mix to ensure dialyzer wouldn't be confused anymore.

* Fixed the spec for fetch_one/0.

* Added missing typing on chat room schema.

* Added new route for picking a giving chat room.

* Refactored seeding logic.

Prior, I had created a custom task for seeding data in Postgres. An issue that rose was that when the Docker container is up, it doesn't mean that the application is running when executing its entrypoint. From research, I saw that the correct practice was to use priv/repo/seed.exs as the task seeding the DB. Furthermore I also found the function to invoke to ensure that the app is running prior to seeding.

* Updated test suite for app home page GET /

* Mounting page with selected chat room if ID in params.

* Leveraging link to avoid dropping socket connection.

* Added default chat room.